Update version to 4.0.0-alpha.3 and refactor font handling
Increases the application version to 4.0.0-alpha.3 to reflect recent updates.

Removes redundant font loading logic and shifts to using font tuples for better management. This simplifies the code and enhances readability while improving font handling across different platforms.

Improves error handling function naming for clarity.
